Module Make.Client

type t

A connected client.

val submit : urgent:bool -> t -> Item.t -> ticket

submit ~urgent t item adds item to the incoming queue. urgent items will be processed before non-urgent ones.

val cancel : t -> ticket -> (unit, [> `Not_queued ]) Stdlib.result

cancel t ticket discards the item from the queue.

val set_rate : t -> float -> unit

set_rate t rate sets the maximum number of jobs that the client can expect to run at once. Clients can submit more jobs than this, and make use of any spare capacity. However, this will determine what happens when multiple clients want to use the extra capacity.

val get_rate : t -> float

get_rate t is the rate previously set by set_rate (or 1.0 if never set).

val client_id : t -> string

val pool_id : t -> string