Module Sqlite_loader.Ty

The database can store different types of resource. It uses strings to identify them, but the rest of the code uses OCaml types. This module maintains a mapping between type names and OCaml types.

type t = ..

val register : string -> t -> unit

register_type name ty registers name as the string representation of ty. A single database may be used to store resources from several libraries, so try to pick a unique type name.

val pp : t Fmt.t